support transferring ManifoldState to gpu
