+++ src/ForwardBackward.jl
@@ -1,48 +1,48 @@
 module ForwardBackward
 
-    using Distributions, LinearAlgebra, ArraysOfArrays
+using Distributions, LinearAlgebra, ArraysOfArrays
 
-    include("maths.jl")
-    include("processes.jl")
-    include("states.jl")
-    include("propagation.jl")
-    include("manifolds.jl")
+include("maths.jl")
+include("processes.jl")
+include("states.jl")
+include("propagation.jl")
+include("manifolds.jl")
 
-    export
-        #Abstract Types
-        Process,
-        DiscreteProcess,
-        State,
-        StateLikelihood,
-        #Processes
-        Deterministic,    
-        BrownianMotion,
-        OrnsteinUhlenbeck,
-        UniformDiscrete,
-        UniformUnmasking,
-        GeneralDiscrete,
-        PiQ,
-        #Likelihoods & States
-        CategoricalLikelihood,
-        GaussianLikelihood,
-        DiscreteState,
-        ContinuousState,
-        #Functions
-        endpoint_conditioned_sample,
-        interpolate,
-        ⊙,
-        forward,
-        backward,
-        forward!,
-        backward!,
-        tensor,
-        sumnorm,
-        stochastic,
-        #Manifolds
-        ManifoldProcess,
-        ManifoldState,
-        perturb!,
-        perturb,
-        expand
+export
+    #Abstract Types
+    Process,
+    DiscreteProcess,
+    State,
+    StateLikelihood,
+    #Processes
+    Deterministic,    
+    BrownianMotion,
+    OrnsteinUhlenbeck,
+    UniformDiscrete,
+    UniformUnmasking,
+    GeneralDiscrete,
+    PiQ,
+    #Likelihoods & States
+    CategoricalLikelihood,
+    GaussianLikelihood,
+    DiscreteState,
+    ContinuousState,
+    #Functions
+    endpoint_conditioned_sample,
+    interpolate,
+    ⊙,
+    forward,
+    backward,
+    forward!,
+    backward!,
+    tensor,
+    sumnorm,
+    stochastic,
+    #Manifolds
+    ManifoldProcess,
+    ManifoldState,
+    perturb!,
+    perturb,
+    expand
     
 end
